requirements/airflow_requirements.txt
pymongo==3.12.1
scrapy==2.5.1
apache-airflow-providers-docker==2.4.0
selenium==3.141.0
